[fresh] Reset useMemoCache when size changes
During development when adding or removing code, it is common for the
cache size to vary between refreshes. Rather than warn like before,
let's simply and safely reset the cache completely whenever its size
changes between renders. In practice, in non-development environments
full invalidation should never happen since the only consumer of uMC is
compiled code – which always emits a static call to uMC with a fixed
size. However, for consistency we continue to reset the cache as well.

ghstack-source-id: 03ae89c2be3b77df85fe3d1c8107940b39ccbdf8
Pull Request resolved: https://github.com/facebook/react/pull/30663